# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  v1.CategorySumResponse:
    properties:
      categoryId:
        description: Category ID
        type: integer
      total:
        description: Total amount in cents for this category
        type: integer
    type: object
...
  v1.CreateOutcomeRequest:
    properties:
      amount:
        description: 'Amount in cents (ex: 1999 for 19.99€)'
        type: integer
      categoryId:
        description: ID of the associated category
        type: integer
      createdAt:
        description: 'Date of the expense (ex: "2026-01-01T00:00:00Z")'
        type: string
      name:
        description: Name of the expense
        type: string
    type: object
...
  v1.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.MonthlySeries:
    properties:
      categories:
        additionalProperties:
          type: integer
        description: Map of categoryId to total amount
        type: object
      month:
        description: Month in YYYY-MM format
        type: string
    type: object
  v1.MonthlyTotalSeries:
    properties:
      month:
        description: Month in YYYY-MM format
        type: string
      total:
        description: Total amount
        type: integer
    type: object
  v1.OutcomeResponse:
    properties:
      amount:
        description: 'Amount in cents (ex: 1999 for 19.99€)'
        type: integer
      categoryId:
        description: ID of the associated category
        type: integer
      createdAt:
        description: 'Date of the expense (ex: "2026-01-01T00:00:00Z")'
        type: string
      id:
        description: ID of the expense
        type: integer
      name:
        description: Name of the expense
        type: string
    type: object
...
  v1.PatchOutcomeRequest:
    properties:
      amount:
        description: 'Amount in cents (optional, ex: 1999 for 19.99€)'
        type: integer
      categoryId:
        description: ID of the associated category (optional)
        type: integer
      createdAt:
        description: 'Date of the expense (optional, ex: "2026-01-01T00:00:00Z")'
        type: string
      name:
        description: Name of the expense (optional)
        type: string
    type: object
  v1.TotalOutcomeResponse:
    properties:
      total:
        description: Total amount in cents
        type: integer
    type: object
...
info:
  contact:
    email: kerhael.me@gmail.com
    name: API Support
  description: API de suivi de comptabilité
  termsOfService: https://example.com/terms/
  title: Accounting API
  version: "1.0"
paths:
...
  /outcomes/:
    get:
      consumes:
      - application/json
      description: Retrieve all outcomes with optional date filtering (defaults to
        current month if not provided)
      parameters:
      - description: Start date filter (ISO 8601 format, defaults to first day of
          current month)
        in: query
        name: from
        type: string
      - description: End date filter (ISO 8601 format, defaults to now)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.OutcomeResponse'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all outcomes
      tags:
      - outcomes
    post:
      consumes:
      - application/json
      description: Create a new outcome
      parameters:
      - description: Outcome payload
        in: body
        name: outcome
        required: true
        schema:
          $ref: '#/definitions/v1.CreateOutcomeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.OutcomeResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an outcome
      tags:
      - outcomes
  /outcomes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an outcome by id
      parameters:
      - description: Outcome ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an outcome
      tags:
      - outcomes
    get:
      consumes:
      - application/json
      description: Retrieve an outcome by id
      parameters:
      - description: Outcome ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.OutcomeResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an outcome
      tags:
      - outcomes
    patch:
      consumes:
      - application/json
      description: Update an outcome
      parameters:
      - description: Outcome ID
        in: path
        name: id
        required: true
        type: integer
      - description: Outcome payload
        in: body
        name: outcome
        required: true
        schema:
          $ref: '#/definitions/v1.PatchOutcomeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.OutcomeResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an outcome
      tags:
      - outcomes
  /outcomes/series-by-category:
    get:
      consumes:
      - application/json
      description: Get the sum of outcomes by category for each month between dates
        (defaults to last 12 months if not provided)
      parameters:
      - description: Start date filter (ISO 8601 format, defaults to 12 months ago)
        in: query
        name: from
        type: string
      - description: End date filter (ISO 8601 format, defaults to now)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/v1.MonthlySeries'
              type: array
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get monthly series of outcomes
      tags:
      - outcomes
  /outcomes/series-total:
    get:
      consumes:
      - application/json
      description: Get the total sum of outcomes for each month between dates (defaults
        to last 12 months if not provided)
      parameters:
      - description: Start date filter (ISO 8601 format, defaults to 12 months ago)
        in: query
        name: from
        type: string
      - description: End date filter (ISO 8601 format, defaults to now)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/v1.MonthlyTotalSeries'
              type: array
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get monthly series of outcomes' total amount
      tags:
      - outcomes
  /outcomes/sums-by-category:
    get:
      consumes:
      - application/json
      description: Get the total amount of outcomes by category between dates (defaults
        to current month if not provided), optionally filtered by category
      parameters:
      - description: Start date filter (ISO 8601 format, defaults to first day of
          current month)
        in: query
        name: from
        type: string
      - description: End date filter (ISO 8601 format, defaults to now)
        in: query
        name: to
        type: string
      - description: Category ID filter
        in: query
        name: categoryId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.CategorySumResponse'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sum of outcomes by category
      tags:
      - outcomes
  /outcomes/total:
    get:
      consumes:
      - application/json
      description: Get the total amount of outcomes between dates (defaults to current
        month if not provided)
      parameters:
      - description: Start date filter (ISO 8601 format, defaults to first day of
          current month)
        in: query
        name: from
        type: string
      - description: End date filter (ISO 8601 format, defaults to now)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TotalOutcomeResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get total of outcomes
      tags:
      - outcomes
...
schemes:
- http
swagger: "2.0"